requirements.txt
flask
PyPDF2
python-docx
pytesseract
pillow
textblob
spacy
flask-cors